+++ modules/whisper/faster_whisper_inference.py
@@ -9,8 +9,8 @@ import whisper
 import gradio as gr
 from argparse import Namespace
 
-from modules.whisper_parameter import *
-from modules.whisper_base import WhisperBase
+from modules.whisper.whisper_parameter import *
+from modules.whisper.whisper_base import WhisperBase
 
 
 class FasterWhisperInference(WhisperBase):
